Added the `ispd::configuration` namespace that represents a dedicated naming space to contain the simulation entities configuration
