+++ .gitignore
@@ -53,3 +53,6 @@ dkms.conf
 
 # Swap files
 .swp
+
+# Clangd
+compile_flags.txt
